MovementScene + Project Structure
Added MovementScene with a very basic script for nonstop 2D movement. Use W and S to move up/down. The speed in the X direction (to the right) is not constant, that will be fixed someday. The scene is divided into a yellow and blue part which stand for the sky and underwater parts. The movement is constrained so that you cannot go backwards and also you cannot fly out of the map. Same for the camera.